requirements.txt
jsonlines
pyppeteer 
python-dotenv
